% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{A Hypothetical Dataset for Illustrative Purpose}
\format{An object of class \code{data.frame} with 10000 rows and 4 columns.}
\usage{
toydata
}
\description{
A dataset containing 4 columns: \code{y} for a continuous outcome,
\code{d} for a binary treatment, \code{x} for a pretreatment covariate,
and \code{z} for an excluded instrument.
}
\keyword{datasets}
